+++ CHANGES.rst
@@ -9,10 +9,13 @@ New Features
 Bug Fixes
 ^^^^^^^^^
 
+- ``bounding_box`` now works with tuple of ``Quantities``. [#331]
+
 - Fix a formula for estimating ``crpix`` in ``to_fits_sip()`` so that ``crpix``
   is near the center of the bounding box. [#337]
 
-- ``bounding_box`` now works with tuple of ``Quantities``. [#331]
+- Allow sub-pixel sampling of the WCS model when computing SIP approximation in
+  ``to_fits_sip()``. [#338]
 
 0.15.0 (2020-11-13)
 -------------------